back/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.0
drf_yasg==1.21.7
openai==0.28.1
pytz==2023.3.post1
django-cors-headers==4.3.0
whitenoise==6.8.2


